// web/src/generated/graphql.tsx
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: any }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A location in a connection that can be used for resuming pagination. */
  Cursor: any;
  /** A point in time as described by the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) standard. May or may not include a timezone. */
  Datetime: any;
};

/** The root query type which gives access points into the data universe. */
export type Query = Node & {
  __typename?: 'Query';
  /** Exposes the root query type nested one level down. This is helpful for Relay 1 which can only query top level fields if they are in a particular form. */
  query: Query;
  /** The root query type must be a `Node` to work well with Relay 1 mutations. This just resolves to `query`. */
  nodeId: Scalars['ID'];
  /** Fetches an object given its globally unique `ID`. */
  node?: Maybe<Node>;
  /** Reads and enables pagination through a set of `SequelizeMeta`. */
  allSequelizeMetas?: Maybe<SequelizeMetasConnection>;
  /** Reads and enables pagination through a set of `Tally`. */
  allTallies?: Maybe<TalliesConnection>;
  /** Reads and enables pagination through a set of `User`. */
  allUsers?: Maybe<UsersConnection>;
  sequelizeMetaByName?: Maybe<SequelizeMeta>;
  tallyById?: Maybe<Tally>;
  userById?: Maybe<User>;
  /** Reads a single `SequelizeMeta` using its globally unique `ID`. */
  sequelizeMeta?: Maybe<SequelizeMeta>;
  /** Reads a single `Tally` using its globally unique `ID`. */
  tally?: Maybe<Tally>;
  /** Reads a single `User` using its globally unique `ID`. */
  user?: Maybe<User>;
};
// ...
export type Node = {
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID'];
};
// ...
export type SequelizeMetasConnection = {
  __typename?: 'SequelizeMetasConnection';
  /** A list of `SequelizeMeta` objects. */
  nodes: Array<Maybe<SequelizeMeta>>;
  /** A list of edges which contains the `SequelizeMeta` and cursor to aid in pagination. */
  edges: Array<SequelizeMetasEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The count of *all* `SequelizeMeta` you could get from the connection. */
  totalCount: Scalars['Int'];
};

export type SequelizeMeta = Node & {
  __typename?: 'SequelizeMeta';
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID'];
  name: Scalars['String'];
};

/** A `SequelizeMeta` edge in the connection. */
export type SequelizeMetasEdge = {
  __typename?: 'SequelizeMetasEdge';
  /** A cursor for use in pagination. */
  cursor?: Maybe<Scalars['Cursor']>;
  /** The `SequelizeMeta` at the end of the edge. */
  node?: Maybe<SequelizeMeta>;
};

/** Information about pagination in a connection. */
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['Cursor']>;
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['Cursor']>;
};
// ...
export type TalliesConnection = {
  __typename?: 'TalliesConnection';
  /** A list of `Tally` objects. */
  nodes: Array<Maybe<Tally>>;
  /** A list of edges which contains the `Tally` and cursor to aid in pagination. */
  edges: Array<TalliesEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The count of *all* `Tally` you could get from the connection. */
  totalCount: Scalars['Int'];
};

export type Tally = Node & {
  __typename?: 'Tally';
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID'];
  id: Scalars['Int'];
  value: Scalars['Int'];
  category?: Maybe<Scalars['String']>;
  note?: Maybe<Scalars['String']>;
  userId: Scalars['Int'];
  createdAt: Scalars['Datetime'];
  updatedAt: Scalars['Datetime'];
  /** Reads a single `User` that is related to this `Tally`. */
  userByUserId?: Maybe<User>;
};

export type User = Node & {
  __typename?: 'User';
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID'];
  id: Scalars['Int'];
  firstName?: Maybe<Scalars['String']>;
  lastName?: Maybe<Scalars['String']>;
  username?: Maybe<Scalars['String']>;
  /** Reads and enables pagination through a set of `Tally`. */
  talliesByUserId: TalliesConnection;
};


export type UserTalliesByUserIdArgs = {
  first?: Maybe<Scalars['Int']>;
  last?: Maybe<Scalars['Int']>;
  offset?: Maybe<Scalars['Int']>;
  before?: Maybe<Scalars['Cursor']>;
  after?: Maybe<Scalars['Cursor']>;
  orderBy?: Maybe<Array<TalliesOrderBy>>;
  condition?: Maybe<TallyCondition>;
};

/** A `Tally` edge in the connection. */
export type TalliesEdge = {
  __typename?: 'TalliesEdge';
  /** A cursor for use in pagination. */
  cursor?: Maybe<Scalars['Cursor']>;
  /** The `Tally` at the end of the edge. */
  node?: Maybe<Tally>;
};
// ...
export type UsersConnection = {
  __typename?: 'UsersConnection';
  /** A list of `User` objects. */
  nodes: Array<Maybe<User>>;
  /** A list of edges which contains the `User` and cursor to aid in pagination. */
  edges: Array<UsersEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The count of *all* `User` you could get from the connection. */
  totalCount: Scalars['Int'];
};
// ...
export type UserFragmentFragment = (
  { __typename?: 'User' }
  & Pick<User, 'firstName' | 'lastName' | 'username'>
);

export type TallyFragmentFragment = (
  { __typename?: 'Tally' }
  & Pick<Tally, 'note' | 'value' | 'category'>
);

export type CurrentUserQueryVariables = Exact<{
  id: Scalars['Int'];
}>;


export type CurrentUserQuery = (
  { __typename?: 'Query' }
  & { userById?: Maybe<(
    { __typename: 'User' }
    & Pick<User, 'id'>
    & { talliesByUserId: (
      { __typename: 'TalliesConnection' }
      & Pick<TalliesConnection, 'totalCount'>
      & { nodes: Array<Maybe<(
        { __typename: 'Tally' }
        & Pick<Tally, 'id' | 'nodeId'>
        & TallyFragmentFragment
      )>> }
    ) }
    & UserFragmentFragment
  )> }
);